refactor: rename and move legacy contract types
